[#841] Logout if user auth token is wrong
This also removes the catch and return stuff. We do not want
to return non okay values as objects because those won't fit
the datamodel we expect in the promise chain later on.

Also: we might want to type the results somehow. We should have
seen that the return value differs via typescript?

closes #841